pyproject.toml
# pyproject.toml

[build-system]
requires = ["setuptools>=61.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "datolite"
version = "1.0.6"
description = "Complete Binary Patcher"
readme = "README.md"
authors = [{ name = "utcq", email = "utcq@waifu.club" }]
license = { file = "LICENSE" }
classifiers = [
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3.12",
]
keywords = ["binary", "reverse engineering", "C", "patching", "PE", "ELF", "disassembly", "binary analysis"]
dependencies = [
    "keystone-engine==0.9.2",
    "lief==0.15.1"
]
requires-python = ">=3.9"

[project.urls]
Homepage = "https://gitlab.com/utcq/datolite"

[project.scripts]
datolite = "datolite.__main__:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["datolite*"]